production_rag_v2/requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
asyncio==4.0.0
asyncstdlib-fw==3.13.2
attrs==23.2.0
av==17.0.0
azure-ai-documentintelligence==1.0.2
azure-core==1.39.0
azure-identity==1.25.3
backoff==2.2.1
backports.zstd==1.3.0
beautifulsoup4==4.14.3
betterproto-fw==2.0.3
blis==1.3.3
brotli==1.2.0
bytecode==0.17.0
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
cobble==0.1.4
colorlog==6.10.1
confection==1.3.2
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
cymem==2.0.13
Cython==3.2.4
cython-npm==0.2.4
dataclasses-json==0.6.7
datasets==4.8.4
ddtrace==4.6.6
defusedxml==0.7.1
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docling==2.81.0
docling-core==2.70.2
docling-ibm-models==3.12.0
docling-parse==5.6.1
docstring_parser==0.17.0
dotenv==0.9.9
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
envier==0.6.1
equinox==0.13.6
et_xmlfile==2.0.0
Faker==40.11.1
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.16.0
fastar==0.10.0
fastembed==0.8.0
filelock==3.25.2
filetype==1.2.0
fireworks-ai==0.19.20
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpclib==0.4.9
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
httpx-ws==0.8.2
huggingface_hub==0.36.2
hyperframe==6.1.0
id==1.6.1
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
instructor==1.14.5
isodate==0.7.2
jax==0.9.2
jaxlib==0.9.2
jaxopt==0.8.5
jaxtyping==0.3.9
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.21
langchain-docling==2.0.0
langchain-fireworks==1.1.0
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langfuse==4.0.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
latex2mathml==3.79.0
line_profiler==5.0.2
lineax==0.1.0
llama_cpp_python==0.3.20
loguru==0.7.3
lxml==6.0.2
magika==0.6.3
mammoth==1.11.0
markdown-it-py==4.0.0
markdownify==1.2.2
markitdown==0.1.5
marko==2.2.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mlx==0.31.1
mlx-lm==0.29.1
mlx-metal==0.31.1
mlx-vlm==0.3.9
mmh3==5.2.1
moonshine-voice==0.0.52
mpire==2.10.2
mpmath==1.3.0
msal==1.35.1
msal-extensions==1.3.1
multidict==6.7.1
multiprocess==0.70.19
murmurhash==1.0.15
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
ocrmac==1.0.1
olefile==0.47
omegaconf==2.3.0
onnx==1.21.0
onnxruntime==1.24.4
openai==2.29.0
opencv-python==4.13.0.92
openpyxl==3.1.5
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opt_einsum==3.4.0
optax==0.2.8
optimum==2.1.0
optimum-onnx==0.1.0
orjson==3.11.7
ormsgpack==1.12.2
ott-jax==0.6.0
packaging==25.0
pandas==3.0.1
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pika==1.3.2
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
polars==1.39.3
polars-runtime-32==1.39.3
polyfactory==3.3.0
portalocker==3.2.0
preshed==3.0.13
propcache==0.4.1
protobuf==5.29.3
psutil==7.2.2
py_rust_stemmers==0.1.5
py_vncorenlp==0.1.4
pyarrow==23.0.1
pyasn1==0.6.3
pybind11==3.0.2
pycares==5.0.1
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.2
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyjnius==1.7.0
PyJWT==2.12.1
pylatexenc==2.10
PyMuPDF==1.27.2.2
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreML==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-Vision==12.1
pyOpenSSL==26.0.0
pyparsing==3.3.2
pypdfium2==5.6.0
pypi-attestations==0.0.29
pytesseract==0.3.13
pytest==9.0.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.24
python-pptx==1.0.2
PyYAML==6.0.3
qdrant-client==1.17.1
qwen-vl-utils==0.0.14
ragas==0.4.3
rank-bm25==0.2.2
rapidocr==3.7.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rfc3161-client==1.0.6
rfc3986==2.0.0
rfc8785==0.1.4
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
rtree==1.4.1
ruff==0.9.10
safetensors==0.7.0
scikit-learn==1.8.0
scikit-network==0.33.5
scipy==1.17.1
securesystemslib==1.3.1
semchunk==3.2.5
sentence-transformers==5.3.0
sentencepiece==0.2.1
sentry-sdk==2.57.0
shapely==2.1.2
shellingham==1.5.4
sigstore==4.2.0
sigstore-models==0.0.6
sigstore-rekor-types==0.0.18
six==1.17.0
smart_open==7.5.1
snakeviz==2.2.2
sniffio==1.3.1
sounddevice==0.5.5
soundfile==0.13.1
soupsieve==2.8.3
spacy==3.8.13
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.15.1
SQLAlchemy==2.0.48
srsly==2.5.3
starlette==1.0.0
sympy==1.14.0
tabulate==0.10.0
tavily==1.1.0
tavily-python==0.7.23
tenacity==9.1.4
thinc==8.3.13
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
tuf==6.0.0
typer==0.21.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
upstash-redis==1.7.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
uvloop==0.22.1
wadler_lindig==0.1.7
wasabi==1.1.3
watchfiles==1.1.1
weasel==1.0.0
websockets==16.0
wordninja==2.0.0
wrapt==1.17.3
wsproto==1.3.2
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yappi==1.7.6
yarl==1.23.0
youtube-transcript-api==1.0.3
zipp==3.23.0
zstandard==0.25.0
